Andres Oct 25 (had PC errors)
was having some PC issues so I couldnt do much but ill get back on it later today.

I changed the amount of turrets that fire because I fount it kind of overwhelming especially considering touch controls.

I think levels prior to the one with the puzzle currently are pretty much at the point where they mainly just need to be technically polished before assets can be added. Let me know what you think about them though, Some things like walls and gates im waiting to see if the asset pack has them and how they look once in before I add but the main skeletons are done. Let me know if you think a level needs more work though.

That being said im going to start diving deeper into Blender or even go to Maya again because it seems like the characters are going to be 100% original so I have to get on that for sure especially since we are having two. Im aiming for us to have a beta version of the game meaning all assets are done and only polish and optimization with real mobille testing is left by the end of November so we can hit that mid/end December release.

(oh and ill probably also dabble with making stuff for the asteroid game :P)